moves sources model from the incidentsModel file to it's own file
